+++ packages/orq-rc/src/orq_ai_sdk/models/runagentop.py
@@ -1808,7 +1808,7 @@ class RunAgentAgentToolInputRunTools(BaseModel):
 
     schema_: Annotated[AgentToolInputRunSchema, pydantic.Field(alias="schema")]
 
-    id: Optional[str] = "01KR16WF5DCEF6Q6CFMG90N4RB"
+    id: Optional[str] = "01KR35JPA7TJ4BRR6WN89X02ST"
 
     description: Optional[str] = None