+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KR16WG3H1PKD84CGDWFNB3YP"
+        "01KR35JPZKPT3A2CJWCJ6HEHTM"
     )
     r"""The unique identifier of the data source"""