+++ packages/orq-rc/src/orq_ai_sdk/models/updateidentityop.py
@@ -194,7 +194,7 @@ class UpdateIdentityResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2026-05-07T12:33:45.946Z")
+    updated: Optional[datetime] = parse_datetime("2026-05-08T06:49:25.887Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")