+++ packages/orq-rc/RELEASES.md
@@ -7639,3 +7639,13 @@ Based on:
 - [python v4.9.0-rc.31] packages/orq-rc
 ### Releases
 - [PyPI v4.9.0-rc.31] https://pypi.org/project/orq-ai-sdk/4.9.0-rc.31 - packages/orq-rc
+
+## 2026-05-08 06:51:48
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.762.0 (2.882.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.9.0-rc.32] packages/orq-rc
+### Releases
+- [PyPI v4.9.0-rc.32] https://pypi.org/project/orq-ai-sdk/4.9.0-rc.32 - packages/orq-rc